requirements.txt
colorama==0.4.6
coverage==7.13.3
iniconfig==2.3.0
packaging==26.0
pluggy==1.6.0
Pygments==2.19.2
PyHamcrest==2.1.0
pytest==9.0.2
pytest-cov==7.0.0
